requirements.txt
python-dotenv
openai
requests
streamlit

#PII
spacy
presidio-analyzer
presidio-anonymizer
faker